+++ Content.Client/VendingMachines/VendingMachineSystem.cs
@@ -1,9 +1,7 @@
-using System.Linq;
 using Content.Shared.VendingMachines;
 using Content.Shared.VendingMachines.Components;
 using Robust.Client.Animations;
 using Robust.Client.GameObjects;
-using Robust.Shared.GameStates;
 
 namespace Content.Client.VendingMachines;
