+++ Content.Client/VendingMachines/VendingMachineVisualsComponent.cs
@@ -1,11 +1,9 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared.VendingMachines.Components;
+namespace Content.Client.VendingMachines;
 
 /// <summary>
 ///
 /// </summary>
-[RegisterComponent, NetworkedComponent]
+[RegisterComponent]
 public sealed partial class VendingMachineVisualsComponent : Component
 {
     /// <summary>
@@ -58,4 +56,22 @@ public sealed partial class VendingMachineVisualsComponent : Component
     /// </summary>
     [DataField("loopDeny")]
     public bool LoopDenyAnimation = true;
+}
+
+public enum VendingMachineVisualLayers : byte
+{
+    /// <summary>
+    /// Off / Broken. The other layers will overlay this if the machine is on.
+    /// </summary>
+    Base,
+
+    /// <summary>
+    /// Normal / Deny / Eject
+    /// </summary>
+    BaseUnshaded,
+
+    /// <summary>
+    /// Screens that are persistent (where the machine is not off or broken)
+    /// </summary>
+    Screen,
 }
