+++ Content.Shared/VendingMachines/Components/VendingMachineContrabandComponent.cs
@@ -1,0 +1,9 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.VendingMachines.Components;
+
+/// <summary>
+///
+/// </summary>
+[RegisterComponent, NetworkedComponent]
+public sealed partial class VendingMachineContrabandComponent : Component;
